Removing fixture imports remaining fixture imports and move fixture constants and utils to seperate files [CodeBuild]
