+++ tests/fixtures/database_connection_fixture_constants.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+BACKEND_ONPREM = 'onprem'
+BACKEND_SAAS = 'saas'
