+++ tests/fixtures/language_container_fixture_constants.py
@@ -1,0 +1,2 @@
+LANGUAGE_ALIAS = "PYTHON3_TE"
+CONTAINER_FILE_NAME = "exasol_transformers_extension_container.tar.gz"
