+++ tests/integration_tests/with_db/udfs/test_sequence_classification_single_text_script.py
@@ -1,18 +1,11 @@
-from tests.fixtures.model_fixture import upload_sequence_classification_model_to_bucketfs
 from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
-
-#debug
-
-
-
-
 
 
 def test_sequence_classification_single_text_script(
         setup_database, db_conn, upload_sequence_classification_model_to_bucketfs):
     bucketfs_conn_name, _ = setup_database
-    n_labels = 3 # negative, neutral, positive
+    n_labels = 3  # negative, neutral, positive
 
     n_rows = 100
     input_data = []
@@ -50,7 +43,7 @@ def test_sequence_classification_single_text_script(
     number_accepted_results = 0
     for i in range(len(result)):
         if (result[i][4] == "positive" and
-                result[i][5] > 0.8): #check if confidence resonably high
+                result[i][5] > 0.8):  # check if confidence resonably high
             number_accepted_results += 1
         elif result[i][5] < 0.2:
             number_accepted_results += 1
