+++ tests/integration_tests/with_db/udfs/test_zero_shot_text_classification_script.py
@@ -1,12 +1,5 @@
-from tests.fixtures.model_fixture import upload_zero_shot_classification_model_to_bucketfs
 from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
-
-# debug
-
-
-
-
 
 
 def test_zero_shot_classification_single_text_script(
@@ -57,7 +50,7 @@ def test_zero_shot_classification_single_text_script(
     number_accepted_results = 0
     for i in range(len(result)):
         if (contains(result[i][5], acceptable_results) and
-                result[i][6] > 0.8):             #check if confidence resonably high
+                result[i][6] > 0.8):  # check if confidence resonably high
             number_accepted_results += 1
         elif result[i][6] < 0.2:
             number_accepted_results += 1
