+++ tests/integration_tests/without_db/udfs/test_filling_mask_udf.py
@@ -9,9 +9,8 @@ from exasol_transformers_extension.udfs.models.filling_mask_udf import \
     FillingMaskUDF
 from tests.integration_tests.without_db.udfs.matcher import Result, ScoreMatcher, RankDTypeMatcher, ShapeMatcher, \
     NoErrorMessageMatcher, NewColumnsEmptyMatcher, ErrorMessageMatcher, RankMonotonicMatcher, ColumnsMatcher
-from tests.utils.parameters import model_params
 from tests.utils.mock_connections import create_mounted_bucketfs_connection
-from tests.fixtures.model_fixture import prepare_filling_mask_model_for_local_bucketfs
+from tests.utils.parameters import model_params
 
 
 class ExaEnvironment:
