+++ tests/integration_tests/without_db/udfs/test_sequence_classification_single_text_udf.py
@@ -5,17 +5,10 @@ import torch
 from exasol_udf_mock_python.connection import Connection
 from exasol_transformers_extension.udfs.models.sequence_classification_single_text_udf import \
     SequenceClassificationSingleTextUDF
-from tests.fixtures.model_fixture import prepare_sequence_classification_model_for_local_bucketfs
 from tests.integration_tests.without_db.udfs.matcher import Result, ShapeMatcher, ColumnsMatcher, NoErrorMessageMatcher, \
     NewColumnsEmptyMatcher, ErrorMessageMatcher
 from tests.utils.parameters import model_params
 from tests.utils.mock_connections import create_mounted_bucketfs_connection
-
-
-
-
-
-
 
 
 class ExaEnvironment:
