+++ tests/integration_tests/without_db/udfs/test_text_generation_udf.py
@@ -7,11 +7,10 @@ from exasol_udf_mock_python.connection import Connection
 
 from exasol_transformers_extension.udfs.models.text_generation_udf import \
     TextGenerationUDF
-from tests.fixtures.model_fixture import prepare_text_generation_model_for_local_bucketfs
 from tests.integration_tests.without_db.udfs.matcher import Result, ShapeMatcher, NewColumnsEmptyMatcher, \
-    ErrorMessageMatcher, ScoreMatcher, ColumnsMatcher, NoErrorMessageMatcher
-from tests.utils.parameters import model_params
+    ErrorMessageMatcher, ColumnsMatcher, NoErrorMessageMatcher
 from tests.utils.mock_connections import create_mounted_bucketfs_connection
+from tests.utils.parameters import model_params
 
 
 class ExaEnvironment:
