+++ tests/integration_tests/without_db/udfs/test_translation_udf.py
@@ -1,4 +1,3 @@
-import torch
 from typing import Dict
 
 import pandas as pd
@@ -10,8 +9,8 @@ from exasol_transformers_extension.udfs.models.translation_udf import \
     TranslationUDF
 from tests.integration_tests.without_db.udfs.matcher import Result, ShapeMatcher, NoErrorMessageMatcher, \
     NewColumnsEmptyMatcher, ErrorMessageMatcher
-from tests.utils.parameters import model_params
 from tests.utils.mock_connections import create_mounted_bucketfs_connection
+from tests.utils.parameters import model_params
 
 
 class ExaEnvironment:
